+++ lib/ex_gen.ex
@@ -22,6 +22,7 @@ defmodule ExGen do
     |> build_collection()
     |> go_to_project_dir()
     |> copy_files()
+    |> generate_ssh_keys()
     |> init_git()
     |> prompt_to_build()
     |> print_end_message()
@@ -95,9 +96,13 @@ defmodule ExGen do
       app: app,
       mod: mod,
       cookie: 48 |> :crypto.strong_rand_bytes() |> Base.encode64(),
-      sup: !!opts[:sup],
+      sup: !!opts[:sup] or !!opts[:ssh],
       rel: !!opts[:rel],
       net: !!opts[:net],
+      push: !!opts[:push],
+      ssh: !!opts[:ssh],
+      ntp: !!opts[:ntp],
+      rtc: !!opts[:rtc],
       contrib: !!opts[:contrib],
       package: !!opts[:package],
       license: opts[:license],
@@ -149,8 +154,14 @@ defmodule ExGen do
     collection =
       []
       |> add_collection(type, true)
-      |> add_collection(:std_sup, !!opts[:sup] and type in [:std, :nerves])
+      |> add_collection(:std_sup, !!opts[:sup] and type in [:std])
+      |> add_collection(
+        :nerves_sup,
+        (!!opts[:sup] or !!opts[:ssh]) and type in [:nerves]
+      )
       |> add_collection(:std_rel, !!opts[:rel] and type in [:std])
+      |> add_collection(:nerves_gen_ssh_keys, !!opts[:push] or !!opts[:ssh])
+      |> add_collection(:nerves_ssh, !!opts[:ssh])
       |> add_collection(:contrib, !!opts[:contrib])
       |> add_collection(:license_mit, opts[:license] == "MIT")
       |> add_collection(:gitsetup, !opts[:no_git])
@@ -171,6 +182,44 @@ defmodule ExGen do
   defp copy_files(%Project{} = project) do
     copy(project.collection, project.assigns)
     unless project.opts[:no_git], do: File.chmod!(".gitsetup", 0o755)
+    project
+  end
+
+  @spec generate_ssh_keys(Project.t()) :: Project.t()
+  defp generate_ssh_keys(%Project{opts: opts} = project) do
+    if opts[:push] || opts[:ssh] do
+      system_dir = "rootfs_overlay/etc/ssh"
+      user_dir = "priv/ssh"
+
+      # Generate target host SSH key.
+      Mix.shell().info([:green, "* generating target host SSH key", :reset])
+      File.mkdir_p!(system_dir)
+
+      :os.cmd(
+        'ssh-keygen -q -t rsa -b 4096 -N "" -f #{system_dir}/ssh_host_rsa_key'
+      )
+
+      # Generate user SSH key.
+      Mix.shell().info([:green, "* generating user SSH key", :reset])
+      File.mkdir_p!(user_dir)
+      :os.cmd('ssh-keygen -q -t rsa -b 4096 -N "" -f #{user_dir}/id_rsa')
+
+      # Get the generated user key.
+      local_key = user_dir |> Path.join("id_rsa.pub") |> File.read!()
+
+      # Get the global user key.
+      global_key =
+        System.user_home!() |> Path.join(".ssh/id_rsa.pub") |> File.read!()
+
+      # Add both user keys to the authorized_keys.
+      user_dir
+      |> Path.join("authorized_keys")
+      |> File.write!(local_key <> global_key)
+
+      # Make the generator executable.
+      File.chmod!("gen-ssh-keys", 0o755)
+    end
+
     project
   end