+++ lib/ex_gen/templates.ex
@@ -35,8 +35,12 @@ defmodule ExGen.Templates do
 
     # Nerves
     "nerves/.formatter.exs" => {:text, ".formatter.exs"},
+    "nerves/.gitignore" => {:eex, ".gitignore"},
+    "nerves/gen-ssh-keys" => {:text, "gen-ssh-keys"},
     "nerves/mix.exs" => {:eex, "mix.exs"},
     "nerves/config/config.exs" => {:eex, "config/config.exs"},
+    "nerves/lib/app_name/application.ex" => {:eex, "lib/:app/application.ex"},
+    "nerves/lib/app_name/ssh_server.ex" => {:eex, "lib/:app/ssh_server.ex"},
     "nerves/rel/plugins/.gitignore" => {:text, "rel/plugins/.gitignore"},
     "nerves/rel/config.exs" => {:eex, "rel/config.exs"},
     "nerves/rel/vm.args" => {:eex, "rel/vm.args"},
@@ -107,7 +111,7 @@ defmodule ExGen.Templates do
       "base/CHANGELOG.md",
       "base/.editorconfig",
       "nerves/.formatter.exs",
-      "base/.gitignore",
+      "nerves/.gitignore",
       "nerves/mix.exs",
       "nerves/config/config.exs",
       "std/lib/app_name.ex",
@@ -117,6 +121,11 @@ defmodule ExGen.Templates do
       "nerves/test/test_helper.exs"
     ]
   end
+
+  # Nerves project optionals
+  defp collection(:nerves_sup), do: ["nerves/lib/app_name/application.ex"]
+  defp collection(:nerves_gen_ssh_keys), do: ["nerves/gen-ssh-keys"]
+  defp collection(:nerves_ssh), do: ["nerves/lib/app_name/ssh_server.ex"]
 
   @doc """
   Adds a collection `name` to the list of collections conditionally.