+++ templates/nerves/config/config.exs
@@ -1,10 +1,13 @@
-use Mix.Config
+use Mix.Config<%= if @push or @ssh do %>
+
+# Adds a rootfs overlay for host SSH key.
+config :nerves, :firmware, rootfs_overlay: "rootfs_overlay"<% else %>
 
 # # Uncomment all or parts of the following to add files to the root
 # # filesystem or modify the firmware archive.
 # config :nerves, :firmware,
 #   rootfs_overlay: "rootfs_overlay",
-#   fwup_conf: "config/fwup.conf"<%= if @net do %>
+#   fwup_conf: "config/fwup.conf"<% end %><%= if @net do %>
 
 # Configures the network.
 config :nerves_network, :default,
@@ -16,10 +19,16 @@ config :nerves_network, :default,
     domain: "example.com"
   ],
   wlan0: [
-    ssid: "SSID",
-    psk: "Secret",
+    ssid: System.get_env("NERVES_WLAN_SSID"),
+    psk: System.get_env("NERVES_WLAN_PSK"),
     key_mgmt: :"WPA-PSK",
     ipv4_address_method: :dhcp
+  ]<% end %><%= if @push do %>
+
+# Configures the SSH key for firmware pushes.
+config :nerves_firmware_ssh,
+  authorized_keys: [
+    "../priv/ssh/id_rsa.pub" |> Path.expand(__DIR__) |> File.read!()
   ]<% end %>
 
 # Use shoehorn to init critical applications before starting ours.